feat: refactored `getTimes`
BREAKING CHANGE: The previous interface was very very clunky. I am not sure this is good but I think it is  better. Now `getTimes` return a dictionary with all the dates instead of a self-instance of suncalc with embedded data. The difference is that you cannot fetch the data anymore as it is was a property (`result.sunrise`). instead you need to use dictionary sintax (`result[.sunrise]`). Even if this is "worse" I think it is a more flexible approach.